Use u32 instead of usize for shards in DataType::ID
Avoids increasing the overall size of a DataType instance (since a usize
can be both 4 or 8 bytes depending on the system).

+++ src/integration.rs
@@ -698,7 +698,7 @@ fn it_auto_increments_columns_with_shards() {
     for i in 0..n {
         // Since we use n shards, each shard should only hold one row
         // (and the ID should always be 1):
-        assert_eq!(result[i][0], DataType::ID(i, 1i64));
+        assert_eq!(result[i][0], DataType::ID(i as u32, 1i64));
     }
 }
